Move the json editor to app utils and clean up css
