+++ src/apputils/index.ts
@@ -6,6 +6,7 @@ export * from './dialog';
 export * from './domutils';
 export * from './hoverbox';
 export * from './iframe';
+export * from './jsoneditor';
 export * from './sanitizer';
 export * from './styling';
 export * from './vdom';
