+++ src/apputils/jsoneditor.css
